refactor: @AuthenticationPrincipal 사용을 위한 인증 구조 개선
